+++ packages/manager/src/components/PrimaryNav/MobileNav.tsx
@@ -1,25 +1,22 @@
 import Backdrop from '@material-ui/core/Backdrop';
 import IconButton from '@material-ui/core/IconButton';
-import ListSubheader from '@material-ui/core/ListSubheader';
 import List from '@material-ui/core/List';
 import ListItem from '@material-ui/core/ListItem';
-import ListItemIcon from '@material-ui/core/ListItemIcon';
 import ListItemText from '@material-ui/core/ListItemText';
 import Collapse from '@material-ui/core/Collapse';
-import KeyboardArrowDown from '@material-ui/icons/KeyboardArrowDown';
+// import KeyboardArrowDown from '@material-ui/icons/KeyboardArrowDown';
 import CloseIcon from '@material-ui/icons/Close';
 import MenuIcon from '@material-ui/icons/Menu';
-import {
-  Menu as ReachMenu,
-  MenuButton,
-  MenuItems,
-  MenuLink,
-  MenuPopover
-} from '@reach/menu-button';
+// import {
+//   Menu as ReachMenu,
+//   MenuButton,
+//   MenuItems,
+//   MenuLink,
+//   MenuPopover
+// } from '@reach/menu-button';
 import * as React from 'react';
 import { Link } from 'react-router-dom';
 import { makeStyles, Theme } from 'src/components/core/styles';
-import RootRef from '../core/RootRef';
 
 interface Props {
   groups: any;
@@ -158,60 +155,100 @@ const useStyles = makeStyles((theme: Theme) => ({
     backgroundColor: 'rgba(50, 54, 60, 0.5)',
     top: 50,
     left: 0,
-    zIndex: 1
+    zIndex: 6
   },
   menuIcon: {
-    color: 'white'
+    color: 'white',
+    fontSize: '1.125rem',
+    height: 50,
+    lineHeight: '20px',
+    '&:hover': {
+      color: 'white'
+    },
+    '& svg': {
+      marginRight: 10
+    }
   },
   menu: {
-    display: 'none'
+    display: 'none',
+    backgroundColor: '#434951',
+    position: 'fixed',
+    left: 0,
+    width: '100vw',
+    zIndex: 1200,
+    '& span': {
+      color: 'white'
+    }
+  },
+  showMenu: {
+    display: 'block'
   }
 }));
+
+type NavGroup =
+  | 'Compute'
+  | 'Network'
+  | 'Storage'
+  | 'Monitors'
+  | 'Marketplace'
+  | 'Help & Support'
+  | 'Community'
+  | 'None';
 
 export const MobileNav: React.FC<Props> = props => {
   const classes = useStyles();
 
   const { groups } = props;
 
-  const ref = React.useRef<HTMLButtonElement>(null);
-
   const [open, setOpen] = React.useState(false);
+  const [openCompute, setOpenCompute] = React.useState(false);
+  const [openNetwork, setOpenNetwork] = React.useState(false);
+  const [openStorage, setOpenStorage] = React.useState(false);
+  const [openMonitors, setOpenMonitors] = React.useState(false);
+
+  const groupMap = {
+    Compute: openCompute,
+    Network: openNetwork,
+    Storage: openStorage,
+    Monitors: openMonitors
+  };
 
   const toggleMenu = () => {
     if (open) {
-      hideMenu();
       setOpen(false);
     } else {
-      showMenu();
       setOpen(true);
     }
-    // console.log(
-    //   document.getElementById('mobile-menu')?.getAttribute('display')
-    // );
-    // if (document.getElementById('mobile-menu').style.display === 'block') {
-    //   hideMenu();
-    //   console.log('hide');
-    // } else {
-    //   showMenu();
-    //   console.log('show');
-    // }
   };
 
-  const showMenu = () => {
-    document.getElementById('mobile-menu')?.style.display = 'block';
-  };
-
-  const hideMenu = () => {
-    document.getElementById('mobile-menu')?.style.display = 'none';
-  };
-
-  const handleClick = () => {
-    setOpen(!open);
+  const handleClick = (group: NavGroup) => {
+    switch (group) {
+      case 'Compute':
+        setOpenCompute(!openCompute);
+        break;
+      case 'Network':
+        setOpenNetwork(!openNetwork);
+        break;
+      case 'Storage':
+        setOpenStorage(!openStorage);
+        break;
+      case 'Monitors':
+        setOpenMonitors(!openMonitors);
+        break;
+      default:
+        return;
+    }
   };
 
   return (
     // <ReachMenu key={window.location.pathname}>
     //   {({ isExpanded }) => {
+    //     // How we are preventing scroll on the body
+    //     if (isExpanded) {
+    //       document.body.classList.add('overflow-hidden');
+    //     } else {
+    //       document.body.classList.remove('overflow-hidden');
+    //     }
     //     return (
     //       <>
     //         <MenuButton
@@ -299,31 +336,71 @@ export const MobileNav: React.FC<Props> = props => {
     //     );
     //   }}
     // </ReachMenu>
+
     <>
-      <IconButton edge="start" onClick={toggleMenu} aria-label="menu">
-        <MenuIcon className={classes.menuIcon} />
+      {/* How we are preventing scroll on the body */}
+      {open
+        ? document.body.classList.add('overflow-hidden')
+        : document.body.classList.remove('overflow-hidden')}
+
+      <Backdrop className={classes.settingsBackdrop} open={open} />
+      <IconButton
+        className={classes.menuIcon}
+        onClick={toggleMenu}
+        aria-label={open ? 'Close menu' : 'Open menu'}
+      >
+        {open ? <CloseIcon /> : <MenuIcon />} Menu
       </IconButton>
-      <List id="mobile-menu" className={classes.menu}>
-        <ListItem button onClick={handleClick}>
-          <ListItemText primary="A" />
-        </ListItem>
-        <Collapse in={open} timeout="auto" unmountOnExit>
-          <List component="div" disablePadding>
-            <ListItem button>
-              <ListItemText primary="Starred" />
-            </ListItem>
-          </List>
-        </Collapse>
-        <ListItem button onClick={handleClick}>
-          <ListItemText primary="B" />
-        </ListItem>
-        <Collapse in={open} timeout="auto" unmountOnExit>
-          <List component="div" disablePadding>
-            <ListItem button>
-              <ListItemText primary="Starred" />
-            </ListItem>
-          </List>
-        </Collapse>
+      <List
+        id="mobile-menu"
+        className={`${classes.menu} ${open && classes.showMenu}`}
+      >
+        {groups.map((thisGroup: any) => {
+          // For each group, filter out hidden links
+          const filteredLinks = thisGroup.links.filter(
+            (thisLink: any) => !thisLink.hide
+          );
+          if (filteredLinks.length === 0) {
+            return null;
+          }
+
+          // Render a singular PrimaryNavLink for links without a group
+          if (thisGroup.group === 'None' && filteredLinks.length === 1) {
+            const link = filteredLinks[0];
+
+            return (
+              <Link to={link.href} onClick={() => setOpen(false)}>
+                <ListItem>
+                  <ListItemText primary={link.display} />
+                </ListItem>
+              </Link>
+            );
+          }
+
+          // Otherwise return a NavGroup (dropdown menu)
+          return (
+            <>
+              <ListItem
+                key={thisGroup.group}
+                button
+                onClick={() => handleClick(thisGroup.group)}
+              >
+                <ListItemText primary={thisGroup.group} />
+              </ListItem>
+              <Collapse in={groupMap[thisGroup.group]}>
+                <List component="div" disablePadding>
+                  {filteredLinks.map((thisLink: any) => (
+                    <ListItem button key={thisLink.group}>
+                      <Link to={thisLink.href} onClick={() => setOpen(false)}>
+                        <ListItemText primary={thisLink.display} />
+                      </Link>
+                    </ListItem>
+                  ))}
+                </List>
+              </Collapse>
+            </>
+          );
+        })}
       </List>
     </>
   );