+++ packages/manager/src/features/Longview/LongviewLanding/LongviewClients.tsx
@@ -22,6 +22,7 @@ import withLongviewClients, {
   Props as LongviewProps
 } from 'src/containers/longview.container';
 import withProfile from 'src/containers/profile.container';
+import useFlags from 'src/hooks/useFlags';
 import { State as StatsState } from 'src/store/longviewStats/longviewStats.reducer';
 import { MapState } from 'src/store/types';
 import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
@@ -33,13 +34,15 @@ import { getUsedStorage } from './Gauges/Storage';
 import DeleteDialog from './LongviewDeleteDialog';
 import LongviewList from './LongviewList';
 import SubscriptionDialog from './SubscriptionDialog';
+import { DocumentTitleSegment } from 'src/components/DocumentTitle';
 
 const useStyles = makeStyles((theme: Theme) => ({
   headingWrapper: {
-    marginBottom: theme.spacing(1)
+    marginBottom: theme.spacing()
   },
   addNew: {
-    marginLeft: 'auto'
+    marginLeft: 'auto',
+    marginRight: 0
   },
   searchbar: {
     width: '100%',
@@ -67,6 +70,17 @@ const useStyles = makeStyles((theme: Theme) => ({
   },
   selectLabel: {
     minWidth: '65px'
+  },
+  cmrSpacing: {
+    [theme.breakpoints.down('md')]: {
+      marginLeft: 0,
+      marginRight: 0
+    }
+  },
+  cmrSpacingAddNew: {
+    [theme.breakpoints.down('md')]: {
+      marginRight: theme.spacing(2)
+    }
   }
 }));
 
@@ -141,6 +155,7 @@ export const LongviewClients: React.FC<CombinedProps> = props => {
   >(false);
 
   const classes = useStyles();
+  const flags = useFlags();
 
   React.useEffect(() => {
     props.getLongviewClients();
@@ -256,7 +271,13 @@ export const LongviewClients: React.FC<CombinedProps> = props => {
 
   return (
     <React.Fragment>
-      <Grid container className={classes.headingWrapper} alignItems="center">
+      <DocumentTitleSegment segment="Clients" />
+      <Grid
+        container
+        className={`${classes.headingWrapper} ${flags.cmr &&
+          classes.cmrSpacing}`}
+        alignItems="center"
+      >
         <Grid item className={`py0 ${classes.searchbar}`}>
           <Search
             placeholder="Filter by client label or hostname"
@@ -279,7 +300,11 @@ export const LongviewClients: React.FC<CombinedProps> = props => {
             hideLabel
           />
         </Grid>
-        <Grid item className={`${classes.addNew} py0`}>
+        <Grid
+          item
+          className={`py0 ${classes.addNew} ${flags.cmr &&
+            classes.cmrSpacingAddNew}`}
+        >
           <AddNewLink
             onClick={handleAddClient}
             label={newClientLoading ? 'Loading...' : 'Add a Client'}