+++ packages/manager/src/features/linodes/LinodesDetail/LinodeActivity/LinodeActivity_CMR.tsx
@@ -17,7 +17,10 @@ const styles = (theme: Theme) =>
   createStyles({
     root: {},
     title: {
-      marginBottom: theme.spacing(2)
+      marginBottom: theme.spacing(2),
+      [theme.breakpoints.down('md')]: {
+        marginLeft: theme.spacing(1)
+      }
     }
   });