+++ packages/manager/src/features/linodes/LinodesDetail/LinodeBackup/LinodeBackup_CMR.tsx
@@ -38,7 +38,6 @@ import TableHead from 'src/components/core/TableHead';
 import TableRow from 'src/components/core/TableRow';
 import Tooltip from 'src/components/core/Tooltip';
 import Typography from 'src/components/core/Typography';
-import { DocumentTitleSegment } from 'src/components/DocumentTitle';
 import Select, { Item } from 'src/components/EnhancedSelect/Select';
 import ErrorState from 'src/components/ErrorState';
 import Notice from 'src/components/Notice';
@@ -79,7 +78,8 @@ type ClassNames =
   | 'scheduleAction'
   | 'chooseTime'
   | 'chooseDay'
-  | 'cancelButton';
+  | 'cancelButton'
+  | 'cmrSpacing';
 
 const styles = (theme: Theme) =>
   createStyles({
@@ -89,7 +89,10 @@ const styles = (theme: Theme) =>
     },
     title: {
       marginTop: theme.spacing(1),
-      marginBottom: theme.spacing(2)
+      marginBottom: theme.spacing(2),
+      [theme.breakpoints.down('md')]: {
+        marginLeft: theme.spacing(1)
+      }
     },
     subTitle: {
       marginBottom: theme.spacing(1)
@@ -121,13 +124,23 @@ const styles = (theme: Theme) =>
       minWidth: 150
     },
     cancelButton: {
-      marginBottom: theme.spacing(1)
+      marginBottom: theme.spacing(1),
+      [theme.breakpoints.down('md')]: {
+        marginLeft: theme.spacing(),
+        marginRight: theme.spacing()
+      }
     },
     snapshotNameField: {
       minWidth: 275
     },
     snapshotGeneralError: {
       minWidth: '100%'
+    },
+    cmrSpacing: {
+      [theme.breakpoints.down('md')]: {
+        marginLeft: theme.spacing(),
+        marginRight: theme.spacing()
+      }
     }
   });
 
@@ -743,7 +756,11 @@ class _LinodeBackup extends React.Component<CombinedProps, State> {
         >
           Cancel Backups
         </Button>
-        <Typography variant="body2" data-qa-cancel-desc>
+        <Typography
+          className={classes.cmrSpacing}
+          variant="body2"
+          data-qa-cancel-desc
+        >
           Please note that when you cancel backups associated with this Linode,
           this will remove all existing backups.
         </Typography>
@@ -765,6 +782,11 @@ class _LinodeBackup extends React.Component<CombinedProps, State> {
           <Typography>
             Cancelling backups associated with this Linode will delete all
             existing backups. Are you sure?
+          </Typography>
+          <Typography style={{ marginTop: 12 }}>
+            <strong>Note: </strong>
+            Once backups for this Linode have been cancelled, you cannot
+            re-enable them for 24 hours.
           </Typography>
         </ConfirmationDialog>
       </React.Fragment>
@@ -794,7 +816,7 @@ class _LinodeBackup extends React.Component<CombinedProps, State> {
   };
 
   render() {
-    const { backupsEnabled, linodeLabel, permissions, type } = this.props;
+    const { backupsEnabled, permissions, type } = this.props;
 
     if (this.props.backups.error) {
       /** @todo remove promise loader and source backups from Redux */
@@ -808,7 +830,6 @@ class _LinodeBackup extends React.Component<CombinedProps, State> {
 
     return (
       <div>
-        <DocumentTitleSegment segment={`${linodeLabel} - Backups`} />
         {backupsEnabled ? (
           <this.Management />
         ) : (