+++ packages/manager/src/features/linodes/LinodesDetail/LinodeNetworking/LinodeNetworking_CMR.tsx
@@ -22,7 +22,6 @@ import {
 import TableBody from 'src/components/core/TableBody';
 import TableHead from 'src/components/core/TableHead';
 import Typography from 'src/components/core/Typography';
-import { DocumentTitleSegment } from 'src/components/DocumentTitle';
 import EntityHeader from 'src/components/EntityHeader';
 import ErrorState from 'src/components/ErrorState';
 import OrderBy from 'src/components/OrderBy';
@@ -179,6 +178,12 @@ class LinodeNetworking extends React.Component<CombinedProps, State> {
 
   componentDidMount() {
     this.refreshIPs();
+  }
+
+  componentDidUpdate(prevProps: CombinedProps) {
+    if (prevProps.linode.id !== this.props.linode.id) {
+      this.refreshIPs();
+    }
   }
 
   openRemoveIPDialog = (IPToRemove: IPAddress) => {
@@ -509,7 +514,6 @@ class LinodeNetworking extends React.Component<CombinedProps, State> {
 
     return (
       <div>
-        <DocumentTitleSegment segment={`${linodeLabel} - Networking`} />
         {readOnly && <LinodePermissionsError />}
         <LinodeNetworkingSummaryPanel
           linodeRegion={linodeRegion}