+++ packages/manager/src/features/linodes/LinodesLanding/index.tsx
@@ -2,9 +2,17 @@ import * as React from 'react';
 import CircleProgress from 'src/components/CircleProgress';
 import { useReduxLoad } from 'src/hooks/useReduxLoad';
 
-const LinodesLanding: React.FC<{}> = _ => {
+interface Props {
+  isDashboard?: boolean;
+}
+
+const LinodesLanding: React.FC<Props> = props => {
   const { _loading } = useReduxLoad(['linodes', 'images']);
-  return _loading ? <CircleProgress /> : <_LinodesLanding />;
+  return _loading ? (
+    <CircleProgress />
+  ) : (
+    <_LinodesLanding isDashboard={props.isDashboard} />
+  );
 };
 
 import _LinodesLanding from './LinodesLanding';